+++ pkg/api/query/v1_test.go
@@ -1201,6 +1201,97 @@ func TestMetadataEndpoints(t *testing.T) {
 	}
 }
 
+// This test uses global Query timeout of 0s, per-request timeout of 10s.
+// This test verifies that lower value of {global Query timeout, per-request timeout} is used for executing queries.
+func TestQueryEndpointsForQueryTimeout(t *testing.T) {
+	lbls := []labels.Labels{
+		{
+			labels.Label{Name: "__name__", Value: "test_metric1"},
+			labels.Label{Name: "foo", Value: "bar"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric1"},
+			labels.Label{Name: "foo", Value: "boo"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric2"},
+			labels.Label{Name: "foo", Value: "boo"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric_replica1"},
+			labels.Label{Name: "foo", Value: "bar"},
+			labels.Label{Name: "replica", Value: "a"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric_replica1"},
+			labels.Label{Name: "foo", Value: "boo"},
+			labels.Label{Name: "replica", Value: "a"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric_replica1"},
+			labels.Label{Name: "foo", Value: "boo"},
+			labels.Label{Name: "replica", Value: "b"},
+		},
+		{
+			labels.Label{Name: "__name__", Value: "test_metric_replica1"},
+			labels.Label{Name: "foo", Value: "boo"},
+			labels.Label{Name: "replica1", Value: "a"},
+		},
+	}
+
+	db, err := e2eutil.NewTSDB()
+	defer func() { testutil.Ok(t, db.Close()) }()
+	testutil.Ok(t, err)
+
+	app := db.Appender(context.Background())
+	for _, lbl := range lbls {
+		for i := int64(0); i < 10; i++ {
+			_, err := app.Append(0, lbl, i*60000, float64(i))
+			testutil.Ok(t, err)
+		}
+	}
+	testutil.Ok(t, app.Commit())
+
+	now := time.Now()
+	globalQueryTimeout := 0 * time.Second
+	perRequestTimeout := "10s"
+
+	qe := promql.NewEngine(promql.EngineOpts{
+		Logger:     nil,
+		Reg:        nil,
+		MaxSamples: 10000,
+		Timeout:    globalQueryTimeout,
+	})
+	api := &QueryAPI{
+		baseAPI: &baseAPI.BaseAPI{
+			Now: func() time.Time { return now },
+		},
+		queryableCreate: query.NewQueryableCreator(nil, nil, store.NewTSDBStore(nil, db, component.Query, nil), 2, globalQueryTimeout),
+		queryEngine: func(int64) *promql.Engine {
+			return qe
+		},
+		gate: gate.New(nil, 4),
+	}
+
+	var tests = []endpointTestCase{
+		{
+			endpoint: api.query,
+			query: url.Values{
+				"query":   []string{"2"},
+				"time":    []string{"123.4"},
+				"timeout": []string{perRequestTimeout},
+			},
+			errType: baseAPI.ErrorTimeout,
+		},
+	}
+
+	for i, test := range tests {
+		if ok := testEndpoint(t, test, fmt.Sprintf("#%d %s", i, test.query.Encode())); !ok {
+			return
+		}
+	}
+}
+
 func TestParseTime(t *testing.T) {
 	ts, err := time.Parse(time.RFC3339Nano, "2015-06-03T13:21:58.555Z")
 	if err != nil {